final-azure-images-cloud-backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
azure-storage-file-share==12.10.0
requests==2.26.0
